Make `Environment.CurrentDirectory` the default "Excel Folder Path" (instead of C:\)
Change the default "Excel Folder Path" to `Environment.CurrentDirectory` when the application starts. Additionally, when the change folder dialog is open, it will automatically expand the folder tree to select the currently selected folder so it is easier for the user to find the folder they are looking for (since it is probably more likely this folder is near the already selected or current working directory folder than it is to the C: root).